requirements.txt
argcomplete==2.0.0
asgiref==3.5.0
boto3==1.20.54
botocore==1.23.54
certifi==2021.10.8
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.0.3
Django==4.0.2
django-s3-sqlite==0.0.4
django-s3-storage==0.13.6
djangorestframework==3.13.1
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==3.3
jmespath==0.10.0
kappa==0.6.0
placebo==0.9.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.27.0
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.62.3
troposphere==3.2.2
urllib3==1.26.8
Werkzeug==1.0.1
wsgi-request-logger==0.4.6
zappa==0.54.1
